+++ src/Collision/collisionshapes.cpp
@@ -20,7 +20,8 @@ center_(_center), width_height_({_width, _height})
 
 /*----- Circle -----*/
 
-CollisionCircle::CollisionCircle(const Vec2& _center, const float _r) : center_(_center), r_(_r)
+CollisionCircle::CollisionCircle(const Vec2& _center, const double _r) : 
+center_(_center), r_(_r)
 {
 
 }
@@ -33,7 +34,8 @@ CollisionCircle::CollisionCircle(const Vec2& _center, const float _r) : center_(
 
 /*----- Capsule -----*/
 
-CollisionCapsule::CollisionCapsule(const Vec2& _a_center, const Vec2& _b_center, float _r) : a_center_(_a_center), b_center_(_b_center), r_(_r)
+CollisionCapsule::CollisionCapsule(const Vec2& _a_center, const Vec2& _b_center, double _r) : 
+a_center_(_a_center), b_center_(_b_center), r_(_r)
 {
 
 }
@@ -46,7 +48,7 @@ CollisionCapsule::CollisionCapsule(const Vec2& _a_center, const Vec2& _b_center,
 
 /*----- Ray -----*/
 
-CollisionRay::CollisionRay(const Vec2& _center, const Vec2& _direction, const float _len) :
+CollisionRay::CollisionRay(const Vec2& _center, const Vec2& _direction, const double _len) :
 p1_(_center), p2_(_direction.x * _len + _center.x, _direction.y * _len + _center.y), direction_normalized_(_direction), len_(_len)
 {
